+++ src/main/java/com/barion/dungeons_enhanced/world/structure/DEDeepCrypt.java
@@ -18,6 +18,7 @@ import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceType;
 import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
 
 public class DEDeepCrypt{
+    public static final String ID = "deep_crypt";
     public static class Capability implements JigsawCapability.IJigsawCapability{
         public static final Capability Instance = new Capability();
         public static final Codec<Capability> CODEC = Codec.unit(Instance);