+++ src/main/java/com/barion/dungeons_enhanced/world/structure/DEDesertTomb.java
@@ -19,13 +19,7 @@ import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
 import java.util.Map;
 
 public class DEDesertTomb{
-    /*private static boolean checkLocation(PieceGeneratorSupplier.Context<NoneFeatureConfiguration> context){
-        if(context.validBiomeOnTop(Heightmap.Types.WORLD_SURFACE_WG)){
-            return DETerrainAnalyzer.isFlatEnough(context.chunkPos(), context.chunkGenerator(), DETerrainAnalyzer.defaultCheckSettings, context.heightAccessor(), context.randomState());
-        }
-
-        return false;
-    }*/
+    public static final String ID = "desert_tomb";
 
     public static class Capability implements JigsawCapability.IJigsawCapability{
         public static final Capability INSTANCE = new Capability();