+++ src/main/java/com/barion/dungeons_enhanced/world/structure/DEMonsterMaze.java
@@ -17,6 +17,7 @@ import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceType;
 import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
 
 public class DEMonsterMaze{
+    public static final String ID = "monster_maze";
 
     public static class Capability implements JigsawCapability.IJigsawCapability{
         public static final Capability INSTANCE = new Capability();