+++ src/main/java/com/barion/dungeons_enhanced/world/structure/DEPillagerCamp.java
@@ -18,6 +18,7 @@ import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceType;
 import net.minecraft.world.level.levelgen.structure.pools.StructureTemplatePool;
 
 public class DEPillagerCamp{
+    public static final String ID = "pillager_camp";
 
     public static class Capability implements JigsawCapability.IJigsawCapability{
         public static final Capability INSTANCE = new Capability();